Address reviewer feedback in PR comment: Created a new folder called 'jos3_utilities.py' in 'jos3_functions' folder and moved these functions there from ' src/pythermalcomfort/shared_functions.py'
